Split ICM42670S sample in 3 differents samples;
- one dedicated to data stream
- one for APEX features
- one for AML
Review KConfig, move log defines to prj.conf
Add VERSION files (recommended by Zephyr)
Review sample.yaml
