+++ samples/sensor/icm42670S/aml_pointing/CMakeLists.txt
