+++ samples/sensor/icm42670S/aml_pointing/arduino_i2c.overlay
