+++ samples/sensor/icm42670S/aml_pointing/arduino_spi.overlay
