+++ samples/sensor/icm42670S/aml_pointing/prj.conf
@@ -6,6 +6,10 @@
 
 CONFIG_SENSOR=y
 
+CONFIG_LOG=y
+CONFIG_LOG_PRINTK=y
+CONFIG_SENSOR_LOG_LEVEL_DBG=n
+
 # Floating point format support
 # Selecting this increases stack size requirements slightly, but increases code size significantly.
 CONFIG_CBPRINTF_FP_SUPPORT=y
