+++ samples/sensor/icm42670S/aml_pointing/sample.yaml
@@ -1,0 +1,9 @@
+sample:
+  name: ICM42670S Sensor AML sample
+tests:
+  sample.sensor.icm42670S:
+    harness: console
+    tags: sensors
+    integration_platforms:
+      - arduino_zero
+      - nrf52dk_nrf52832
