+++ samples/sensor/icm42670S/apex/prj.conf
@@ -1,0 +1,15 @@
+#
+# Copyright (c) 2019 TDK Invensense
+#
+# SPDX-License-Identifier: Apache-2.0
+#
+
+CONFIG_SENSOR=y
+
+CONFIG_LOG=y
+CONFIG_LOG_PRINTK=y
+CONFIG_SENSOR_LOG_LEVEL_DBG=n
+
+# Floating point format support
+# Selecting this increases stack size requirements slightly, but increases code size significantly.
+CONFIG_CBPRINTF_FP_SUPPORT=y
