+++ samples/sensor/icm42670S/data_stream/CMakeLists.txt
@@ -1,0 +1,12 @@
+#
+# Copyright (c) 2019 TDK Invensense
+#
+# SPDX-License-Identifier: Apache-2.0
+#
+
+cmake_minimum_required(VERSION 3.20.0)
+find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
+project(icm42670S)
+
+FILE(GLOB app_sources src/*.c)
+target_sources(app PRIVATE ${app_sources})
