+++ samples/sensor/icm42670S/data_stream/arduino_i2c.overlay
@@ -1,0 +1,29 @@
+/*
+ * Copyright (c) 2020, TDK Invensense
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/*
+ * Example configuration of a ICM42670S device on an Arduino I2C bus.
+ *
+ * Device address 0x68 is assumed. Your device may have a different
+ * address; check your device documentation if unsure.
+ */
+&arduino_i2c {
+	status = "okay";
+	icm42670S@68 {
+		compatible = "invensense,icm42670S";
+		reg = <0x68>;
+		int-gpios = <&arduino_header 8 GPIO_ACTIVE_HIGH>; /* D2 */
+	};
+};
+
+/*
+ * Increase native UART speed to report all IMU data at 100Hz.
+ */
+&uart0 {
+	compatible = "nordic,nrf-uarte";
+	status = "okay";
+	current-speed = <1000000>;
+};
