+++ samples/sensor/icm42670S/data_stream/arduino_spi.overlay
@@ -1,0 +1,26 @@
+/*
+ * Copyright (c) 2020, TDK Invensense
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/* Example configuration of a ICM42670S device on an Arduino SPI bus. */
+&arduino_spi {
+	status = "okay";
+	cs-gpios = <&arduino_header 16 GPIO_ACTIVE_LOW>; /* D10 */
+	icm42670S@0 {
+		compatible = "invensense,icm42670S";
+		reg = <0>;
+		spi-max-frequency = <1000000>; /* conservatively set to 1MHz */
+		int-gpios = <&arduino_header 8 GPIO_ACTIVE_HIGH>; /* D2 */
+	};
+};
+
+/*
+ * Increase native UART speed to report all IMU data at 100Hz.
+ */
+&uart0 {
+	compatible = "nordic,nrf-uarte";
+	status = "okay";
+	current-speed = <1000000>;
+};
