+++ samples/sensor/icm42670S/data_stream/prj.conf
@@ -1,0 +1,21 @@
+#
+# Copyright (c) 2019 TDK Invensense
+#
+# SPDX-License-Identifier: Apache-2.0
+#
+
+CONFIG_SENSOR=y
+
+CONFIG_LOG=y
+CONFIG_LOG_PRINTK=y
+CONFIG_SENSOR_LOG_LEVEL_DBG=n
+
+# Floating point format support
+# Selecting this increases stack size requirements slightly, but increases code size significantly.
+CONFIG_CBPRINTF_FP_SUPPORT=y
+
+# Buffer size for console. The defaults are optimized to save RAM
+CONFIG_CONSOLE_SUBSYS=y
+CONFIG_CONSOLE_GETCHAR=y
+CONFIG_CONSOLE_GETCHAR_BUFSIZE=16
+CONFIG_CONSOLE_PUTCHAR_BUFSIZE=512
