--- samples/sensor/icm42670S/sample.yaml
@@ -1,41 +1,0 @@
-sample:
-  name: ICM42670S Sensor sample
-tests:
-  sample.sensor.icm42670S:
-    harness: console
-    tags: sensors
-    platform_allow:
-      - arduino_zero
-      - nrf52dk_nrf52832
-    integration_platforms:
-      - arduino_zero
-      - nrf52dk_nrf52832
-    harness_config:
-      type: one_line
-      regex:
-        - "accel: (.*); gyro: (.*);"
-      fixture: fixture_i2c_icm42670S
-  sample.sensor.icm42670S.spi:
-    harness: console
-    tags: sensors
-    depends_on:
-      - spi
-      - icm42670S
-    extra_args: "DTC_OVERLAY_FILE=arduino_spi.overlay"
-    harness_config:
-      type: one_line
-      regex:
-        - "accel: (.*); gyro: (.*);"
-      fixture: fixture_spi_icm42670S
-   sample.sensor.icm42670S.i2c:
-    harness: console
-    tags: sensors
-    depends_on:
-      - i2c
-      - icm42670S
-    extra_args: "DTC_OVERLAY_FILE=arduino_i2c.overlay"
-    harness_config:
-      type: one_line
-      regex:
-        - "accel: (.*); gyro: (.*);"
-      fixture: fixture_i2c_icm42670S
